refactor: move files Certificates.tsx, CheckBox.tsx, Months.tsx, Years.tsx to certificates folder
- From the src/components directory, move
  1. Certificates.tsx to certificates/certificates.tsx
  2. CheckBox.tsx to certificates/check-box.tsx
  3. Months.tsx to certificates/months.tsx
  4. Years.tsx to certificates/years.tsx
- Within the file certificates.tsx, update the paths to the following
  components:
    1. CheckBox
    2. Months
    3. Years
    4. Details
- Update the path to the Certificates component for the Resume and
  Preview components
